# packages/relayer/openapi.yaml
openapi: 3.0.3
info:
  title: 1Prime Relayer Service API
  description: |
    A 1inch Fusion+ compatible relayer service that facilitates cross-chain atomic swaps between EVM chains (Ethereum, Base, BSC, Polygon, Arbitrum) and NEAR Protocol.

    This implementation follows the 1inch Fusion+ whitepaper architecture for secure cross-chain atomic swaps.

    ## Features
    - Partial Fill Support with Merkle tree-based N+1 secret management  
    - Cross-chain Escrow Verification
    - Real-time WebSocket updates
    - Modular chain support for EVM and NEAR

    ## Supported Chain Pairs
    - Ethereum ↔ NEAR
    - Base ↔ NEAR
    - BSC ↔ NEAR
    - Polygon ↔ NEAR
    - Arbitrum ↔ NEAR
  version: 1.0.0
  contact:
    name: 1Prime Team
    url: https://github.com/unite-defi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://1prime-relayer.up.railway.app/api/v1
    description: Production server (Railway)
paths:
...
  /orders:
    get:
      summary: Get Active Orders
      description: Retrieve all currently active orders
      operationId: getActiveOrders
      tags:
        - Orders
      responses:
        "200":
          description: Active orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                success: true
                data:
                  - orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                    maker: "0xeb0d8736Cc2c47882f112507cc8A3355d37D2413"
                    userSrcAddress: "0xeb0d8736Cc2c47882f112507cc8A3355d37D2413"
                    userDstAddress: "user.near"
                    sourceChain: "ethereum"
                    destinationChain: "near-testnet"
                    sourceToken: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
                    destinationToken: "3e2210e1184b45b64c8a434c0a7e7b23cc04ea7eb7a6c3c32520d03d4afcb8af"
                    sourceAmount: "10000000"
                    destinationAmount: "10000000"
                    phase: "claimed"
                    assignedResolver: "0xdef456789abcdef123456789abcdef123456789a"
                timestamp: 1640995200000
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/prepare:
    post:
      summary: Generate Unsigned Order
      description: |
        Generate an unsigned Fusion+ order for frontend signing (Steps 3+4+5 of the cross-chain flow).

        This endpoint uses the 1inch SDK internally to create a properly structured CrossChainOrder
        with HashLock and TimeLocks. The frontend then signs this order.

        Returns the orderHash along with success/message fields for the prepared order.
        The full order details are stored in the database and can be retrieved during submission.
      operationId: generateOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOrderRequest"
            example:
              userSrcAddress: "0xeb0d8736Cc2c47882f112507cc8A3355d37D2413"
              userDstAddress: "papasins.near"
              amount: "10000000"
              fromToken: "usdc"
              toToken: "usdc"
              fromChain: "11155111"
              toChain: "398"
              secretHash: "0xa82aec019867b7307551dc397acde18b541e742fa1a4e53df4ce3b02d462f524"
      responses:
        "200":
          description: Unsigned order generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                success: true
                data:
                  orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                  success: true
                  message: "Order prepared successfully. Use this orderHash with your signature to submit."
                timestamp: 1640995200000
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/submit:
    post:
      summary: Submit Signed Order
      description: |
        Submit a signed Fusion+ order to the relayer (Step 6 of the cross-chain flow).

        The order must be signed by the frontend using the unsigned order from /orders/prepare.
      operationId: submitSignedOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitSignedOrderRequest"
            example:
              orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
              signature: "0x304cf8f9e7a6dba22d1ee4ad59e9e8a35bb2e7d4c1a9f1a8b2c3d4e5f6789012345a1b2c3d4e5f6789012345a1b2c3d4e5f6789012345a1b2c3d4e5f67890123456"
      responses:
        "201":
          description: Signed order submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                success: true
                data:
                  orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                  phase: "submitted"
                  assignedResolver: null
                  sourceEscrow: null
                  destinationEscrow: null

                  secret: null
                  timelock:
                    phase: "submitted"
                    orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                    startTime: 1640995200000
                    endTime: 1640995500000
                    isActive: true
                    nextPhase: "claimed"
                  isCompleted: false
                  events:
                    - type: "order_created"
                      timestamp: 1640995200000
                      data:
                        userSrcAddress: "0xeb0d8736Cc2c47882f112507cc8A3355d37D2413"
                        userDstAddress: "papasins.near"
                        sourceChain: "11155111"
                        destinationChain: "398"
                timestamp: 1640995200000
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{hash}/status:
    get:
      summary: Get Order Status
      description: Retrieve the current status and details of a specific order
      operationId: getOrderStatus
      tags:
        - Orders
      parameters:
        - name: hash
          in: path
          required: true
          description: Unique hash identifier of the order
          schema:
            type: string
            pattern: "^(0x)?[a-fA-F0-9]{40,70}$"
          example: "0x123abc..."
      responses:
        "200":
          description: Order status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                success: true
                data:
                  orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                  phase: "src_escrow_deployed"
                  assignedResolver: "0xdef456789abcdef123456789abcdef123456789a"
                  sourceEscrow:
                    orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                    chain: "ethereum"
                    contractAddress: "0x128ce802AB730FbB360b784CA8C16dD73147649c"
                    designated: "0xdef456789abcdef123456789abcdef123456789a"
                    amount: "10000000"
                    isCreated: true
                  destinationEscrow:
                    orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                    chain: "near-testnet"
                    contractAddress: "escrow-dst-template.testnet"
                    designated: "0xdef456789abcdef123456789abcdef123456789a"
                    amount: "10000000"
                    isCreated: true

                  isCompleted: false
                timestamp: 1640995200000
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{hash}/verify-escrows:
    get:
      summary: Verify Escrows Safety
      description: |
        Perform independent on-chain verification that both source and destination escrows 
        are safe for secret revelation. This is a critical security endpoint that the frontend
        MUST call before revealing any secrets.

        **Security Features:**
        - Verifies contract existence on both chains
        - Checks escrow parameters (amount, token, secret hash)
        - Validates escrow funding status
        - Detects if escrows have already been withdrawn or cancelled
        - Confirms escrow state matches expected conditions

        **NEAR Verification:** Uses NEAR API to call `get_escrow_info()` and verify state
        **EVM Verification:** Uses 1inch contract events to detect withdrawals/cancellations
      operationId: verifyEscrowSafety
      tags:
        - Security
      parameters:
        - name: hash
          in: path
          required: true
          description: Unique hash identifier of the order
          schema:
            type: string
            pattern: "^(0x)?[a-fA-F0-9]{40,70}$"
          example: "0x123abc..."
      responses:
        "200":
          description: Escrow verification completed (check safe field for result)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowVerificationResponse"
              examples:
                safe:
                  summary: Escrows are safe
                  value:
                    success: true
                    data:
                      safe: true
                      orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                      verification:
                        srcEscrowVerified: true
                        dstEscrowVerified: true
                        srcEscrowDetails:
                          escrowAddress: "0x128ce802AB730FbB360b784CA8C16dD73147649c"
                          chainName: "ethereum"
                          contractExists: true
                          amountMatches: true
                          tokenMatches: true
                          secretHashMatches: true
                          properlyFunded: true
                        dstEscrowDetails:
                          escrowAddress: "escrow-dst-template.testnet"
                          chainName: "near-testnet"
                          contractExists: true
                          amountMatches: true
                          tokenMatches: true
                          secretHashMatches: true
                          properlyFunded: true
                      message: "Safe to reveal secret"
                    timestamp: 1640995200000
                unsafe:
                  summary: Escrows are not safe
                  value:
                    success: false
                    data:
                      safe: false
                      orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                      verification:
                        srcEscrowVerified: false
                        dstEscrowVerified: true
                        issues:
                          - "❌ Escrow already withdrawn! Secret revealed: 0xa82aec019867b7307551dc397acde18b541e742fa1a4e53df4ce3b02d462f524"
                          - "❌ Insufficient USDC balance: expected 10000000, got 0"
                      message: "NOT safe to reveal secret"
                    timestamp: 1640995200000
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{hash}/claim:
    post:
      summary: Claim Order
      description: |
        Claim an order for resolver processing (Step 7 of the cross-chain flow).

        Any resolver can claim unclaimed orders. This transitions the order
        to "claimed" phase and assigns it to the resolver address.
      operationId: claimOrder
      tags:
        - Resolvers
      parameters:
        - name: hash
          in: path
          required: true
          description: Unique hash identifier of the order
          schema:
            type: string
            pattern: "^(0x)?[a-fA-F0-9]{40,70}$"
          example: "0x123abc..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimOrderRequest"
            example:
              resolverAddress: "0xdef456789abcdef123456789abcdef123456789a"
      responses:
        "200":
          description: Order claimed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                success: true
                data:
                  claimed: true
                  orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                  resolverAddress: "0xdef..."
                timestamp: 1640995200000
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{hash}/escrow-deployed:
    post:
      summary: Confirm Escrow Deployment
      description: |
        Confirm that an escrow has been deployed (Steps 7.1 & 9.1 of the cross-chain flow).

        The assigned resolver calls this endpoint after deploying source or destination escrows
        to update the order state and provide deployment details.
      operationId: confirmEscrowDeployment
      tags:
        - Resolvers
      parameters:
        - name: hash
          in: path
          required: true
          description: Unique hash identifier of the order
          schema:
            type: string
            pattern: "^(0x)?[a-fA-F0-9]{40,70}$"
          example: "0x123abc..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscrowDeploymentConfirmation"
            example:
              escrowType: "src"
              escrowAddress: "0x789abcdef123456789abcdef123456789abcdef12"
              transactionHash: "0xdef456789abcdef123456789abcdef123456789abcdef123456789abcdef123456"
              blockNumber: 18500000
              resolverAddress: "0xabc123456789abcdef123456789abcdef123456789"
      responses:
        "200":
          description: Escrow deployment confirmed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                success: true
                data:
                  confirmed: true
                  orderHash: "0xa3577a7705e04caebabb32e62a231660b7f5556af20a74220d1608c2a9b62d9c"
                  escrowType: "src"
                  escrowAddress: "0x789..."
                timestamp: 1640995200000
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        data:
          type: object
          description: Response data (present on success)
        error:
          type: string
          description: Error message (present on failure)
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the response
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
          example:
            success: false
            error: "Invalid request parameters"
            timestamp: 1640995200000
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
          example:
            success: false
            error: "Order not found"
            timestamp: 1640995200000
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
          example:
            success: false
            error: "Internal server error"
            timestamp: 1640995200000
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional JWT token for authenticated endpoints
...